// autogpt_platform/frontend/src/app/api/__generated__/models/graph.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { Node } from "./node";
import type { Link } from "./link";
import type { GraphForkedFromId } from "./graphForkedFromId";
import type { GraphForkedFromVersion } from "./graphForkedFromVersion";
import type { BaseGraphInput } from "./baseGraphInput";

export interface Graph {
  id?: string;
  version?: number;
  is_active?: boolean;
  name: string;
  description: string;
  nodes?: Node[];
  links?: Link[];
  forked_from_id?: GraphForkedFromId;
  forked_from_version?: GraphForkedFromVersion;
  sub_graphs?: BaseGraphInput[];
}